コメントのbefore actionを共通化
