+++ app/controllers/reports/comments_controller.rb
@@ -1,11 +1,9 @@
 # frozen_string_literal: true
 
-require_relative '../comments/comments_module'
+require_relative '../concerns/comments_module'
 
 class Reports::CommentsController < ApplicationController
   include Comments
-
-  before_action :set_commentable
 
   private
